Add `reported_by` tag to error backend errors
We get support tickets about incomplete error reporting, which we
assume is due to the use of the error backend. Often customers do not
know whether the error was raised from an uninstrumented context, or
whether they're using a version of the integration where the error
backend is enabled by default.

This should simplify the support workflow for it going forward, by
giving us that information in the error sample itself. We can also
ask customers to update to the latest version and see if the affected
samples have this tag.

This follows the format that we use in [the Ruby integration][ruby]

[ruby]: https://github.com/appsignal/appsignal-ruby/blob/6201f5dbe662e22e682a8a324b7e996e979ef09f/lib/appsignal/integrations/puma.rb#L21